Code cleanup, added convenience methods to simplify ceremonial code, and provided better naming scope for RepoDb extensions, etc.
Renamed RepoDb extension methods to scope them more appropriately as specific to this GraphQL extension pack.  If Relay Cursor paginated Batch Querying is ever added to the RepoDb core, this these methods wills till remain as a proxy specific to the GraphQL integration --providing necessary mapping between CursorSlice classes in the core the those in the HotChocolate PreProcessing extensions.